+++ CHANGELOG.md
@@ -36,3 +36,6 @@
 * Added the AfterPay localization for Netherlands, German and French language.
 * Klarna Pay Later : Remove fields gender and birthday from checkout form and customer object.
 * Replaced locally stored ginger-php library on composer library installer.
+* Added order lines for Klarna Pay Later payment method
+* Added order lines for AfterPay payment method
+